Move scaler-related APIs into dedicated package
