+++ pkg/scaler/scaler.go
@@ -9,6 +9,7 @@ import (
 	"fmt"
 	"time"
 
+	"github.com/gardener/dependency-watchdog/pkg/scaler/api"
 	v1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/api/meta"
 	utilruntime "k8s.io/apimachinery/pkg/util/runtime"
@@ -34,7 +35,7 @@ func NewController(clientset kubernetes.Interface,
 	scalesGetter scale.ScalesGetter,
 	sharedInformerFactory informers.SharedInformerFactory,
 	gardenerInformerFactory gardenerinformers.SharedInformerFactory,
-	probeDependantsList *ProbeDependantsList,
+	probeDependantsList *api.ProbeDependantsList,
 	stopCh <-chan struct{}) *Controller {
 
 	c := &Controller{
@@ -239,7 +240,7 @@ func (c *Controller) processNamespace(key string) error {
 	for i := range c.probeDependantsList.Probes {
 		probeDeps := &c.probeDependantsList.Probes[i]
 
-		go func(ns string, pd *ProbeDependants) {
+		go func(ns string, pd *api.ProbeDependants) {
 			p := &prober{
 				namespace:         ns,
 				mapper:            c.mapper,
@@ -286,7 +287,7 @@ func (c *Controller) processNamespace(key string) error {
 	return nil
 }
 
-func (c *Controller) getKey(ns string, probeDeps *ProbeDependants) string {
+func (c *Controller) getKey(ns string, probeDeps *api.ProbeDependants) string {
 	return ns + "/" + probeDeps.Name
 }
 
@@ -337,7 +338,7 @@ func (c *Controller) deleteProber(key string) {
 	delete(c.probers, key)
 }
 
-func (c *Controller) newContext(ns string, probeDeps *ProbeDependants) (context.Context, context.CancelFunc) {
+func (c *Controller) newContext(ns string, probeDeps *api.ProbeDependants) (context.Context, context.CancelFunc) {
 	key := c.getKey(ns, probeDeps)
 
 	ctx, cancelFn := context.WithCancel(context.Background())
