+++ pkg/scaler/types.go
@@ -8,16 +8,16 @@ import (
 	"sync"
 
 	"github.com/gardener/dependency-watchdog/pkg/multicontext"
+	"github.com/gardener/dependency-watchdog/pkg/scaler/api"
 	gardnerinformer "github.com/gardener/gardener/pkg/client/extensions/informers/externalversions"
 	gardenerlisterv1alpha1 "github.com/gardener/gardener/pkg/client/extensions/listers/extensions/v1alpha1"
 	"github.com/prometheus/client_golang/prometheus"
-	autoscaling "k8s.io/api/autoscaling/v1"
 	apimeta "k8s.io/apimachinery/pkg/api/meta"
 	"k8s.io/client-go/informers"
 	"k8s.io/client-go/kubernetes"
 	listerappsv1 "k8s.io/client-go/listers/apps/v1"
 	listerv1 "k8s.io/client-go/listers/core/v1"
-	scale "k8s.io/client-go/scale"
+	"k8s.io/client-go/scale"
 	"k8s.io/client-go/tools/cache"
 	"k8s.io/client-go/util/workqueue"
 	componentbaseconfig "k8s.io/component-base/config/v1alpha1"
@@ -41,50 +41,12 @@ type Controller struct {
 	hasClustersSynced      cache.InformerSynced
 	hasDeploymentsSynced   cache.InformerSynced
 	stopCh                 <-chan struct{}
-	probeDependantsList    *ProbeDependantsList
+	probeDependantsList    *api.ProbeDependantsList
 	probers                map[string]*prober // the key is <namespace>/<probeDependents.Name>
 	mux                    sync.Mutex
 	*multicontext.Multicontext
 	// LeaderElection defines the configuration of leader election client.
 	LeaderElection componentbaseconfig.LeaderElectionConfiguration
-}
-
-// ProbeDependantsList holds a list of probes (internal and external) and their corresponding
-// dependant Scales. If the external probe fails and the internal probe still succeeds, then the
-// corresponding dependant Scales are scaled down to `zero`. They are scaled back to their
-// original scale when the external probe succeeds again.
-type ProbeDependantsList struct {
-	Probes    []ProbeDependants `json:"probes"`
-	Namespace string            `json:"namespace"`
-}
-
-// ProbeDependants struct captures the details about a probe and its dependant scale sub-resources.
-type ProbeDependants struct {
-	Name            string                   `json:"name"`
-	Probe           *ProbeConfig             `json:"probe"`
-	DependantScales []*DependantScaleDetails `json:"dependantScales"`
-}
-
-// ProbeConfig struct captures the details for probing a Kubernetes apiserver.
-type ProbeConfig struct {
-	External            *ProbeDetails `json:"external,omitempty"`
-	Internal            *ProbeDetails `json:"internal,omitempty"`
-	InitialDelaySeconds *int32        `json:"initialDelaySeconds,omitempty"`
-	TimeoutSeconds      *int32        `json:"timeoutSeconds,omitempty"`
-	PeriodSeconds       *int32        `json:"periodSeconds,omitempty"`
-	SuccessThreshold    *int32        `json:"successThreshold,omitempty"`
-	FailureThreshold    *int32        `json:"failureThreshold,omitempty"`
-}
-
-// ProbeDetails captures the kubeconfig secret details to probe a Kubernetes apiserver.
-type ProbeDetails struct {
-	KubeconfigSecretName string `json:"kubeconfigSecretName"`
-}
-
-// DependantScaleDetails has the details about the dependant scale sub-resource.
-type DependantScaleDetails struct {
-	ScaleRef autoscaling.CrossVersionObjectReference `json:"scaleRef"`
-	Replicas *int32                                  `json:"replicas"`
 }
 
 const (
